+++ Gui/consts.py
@@ -1,0 +1,6 @@
+CONST_NOT_ALL_FIELDS_FILLED = "Please fill all fields"
+
+sensor_type_list = ["bme280","bmp180", "bmp280","ds18b20", "dht22", "hpm", "htu21d",
+                    "laerm", "pms1003", "pms3003", "pms5003", "pms7003", "ppd42ns", 
+                    "radiation_sbm-19","radiation_sbm-20","radiation_si22g","sds011",
+                    "sdc30","sht15","sht30","sht31","sht35","sht85","sps30"]
